backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
firebase-admin
python-dotenv
pytest
pytest-flask